# apps/frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "DashboardSummary": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "filters": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Filters",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "visibility": {
            "enum": [
              "private",
              "published"
            ],
            "title": "Visibility",
            "type": "string"
          },
          "widgets": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Widgets",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "widgets",
          "filters",
          "visibility",
          "created_at",
          "updated_at"
        ],
        "title": "DashboardSummary",
        "type": "object"
      },
      "DashboardWrite": {
        "properties": {
          "filters": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "maxItems": 20,
            "title": "Filters",
            "type": "array"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "widgets": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "maxItems": 50,
            "title": "Widgets",
            "type": "array"
          }
        },
        "required": [
          "name"
        ],
        "title": "DashboardWrite",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Data BI With LoL",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/dashboards": {
      "get": {
        "operationId": "dashboards_api_v1_dashboards_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DashboardSummary"
                  },
                  "title": "Response Dashboards Api V1 Dashboards Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Dashboards"
      },
      "post": {
        "operationId": "create_dashboard_api_v1_dashboards_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Dashboard"
      }
    },
    "/api/v1/dashboards/{dashboard_id}": {
      "get": {
        "operationId": "dashboard_api_v1_dashboards__dashboard_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "dashboard_id",
            "required": true,
            "schema": {
              "title": "Dashboard Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dashboard"
      },
      "put": {
        "operationId": "edit_dashboard_api_v1_dashboards__dashboard_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "dashboard_id",
            "required": true,
            "schema": {
              "title": "Dashboard Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edit Dashboard"
      }
    },
    "/api/v1/dashboards/{dashboard_id}/clone": {
      "post": {
        "operationId": "duplicate_dashboard_api_v1_dashboards__dashboard_id__clone_post",
        "parameters": [
          {
            "in": "path",
            "name": "dashboard_id",
            "required": true,
            "schema": {
              "title": "Dashboard Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Duplicate Dashboard"
      }
    },
    "/api/v1/dashboards/{dashboard_id}/publish": {
      "post": {
        "operationId": "publish_dashboard_api_v1_dashboards__dashboard_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "dashboard_id",
            "required": true,
            "schema": {
              "title": "Dashboard Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "published",
            "required": false,
            "schema": {
              "default": true,
              "title": "Published",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Dashboard"
      }
    },
...
    "/api/v1/datasets/upload": {
      "post": {
        "operationId": "upload_dataset_api_v1_datasets_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_dataset_api_v1_datasets_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Dataset"
      }
    },
    "/api/v1/datasets/{dataset_id}/chart": {
      "post": {
        "operationId": "dataset_chart_api_v1_datasets__dataset_id__chart_post",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetChartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetChartResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dataset Chart"
      }
    },
    "/api/v1/datasets/{dataset_id}/preview": {
      "get": {
        "operationId": "dataset_preview_api_v1_datasets__dataset_id__preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Dataset Preview Api V1 Datasets  Dataset Id  Preview Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dataset Preview"
      }
    },
    "/api/v1/datasets/{dataset_id}/profile": {
      "get": {
        "operationId": "dataset_profile_api_v1_datasets__dataset_id__profile_get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dataset Profile"
      }
    },
...
  }
}